BETA: Continuing with [5214] and [5217] Quest class related documentation and formatting fixes, this time by '''jurchiks'''. 	* Reviewed and corrected by me.
'''Note:''' To contributors, checking such long patches take time, and we (developers) are responsible of what gets committed, we cannot and will not commit stuff blindly just because it doesn't change code or it won't break anything. Pushing the development team for commit doesn't help at all, keep that in mind.